Add files via upload
This is a ML model that predicts whether a hotel booking is likely to be cancelled or not with an accuracy of 99.3% . The model takes into account various factors such as Hotel_type, Meal, Seasonality and many other .The problem we're trying to solve is whether the hotel either cancels the booking or not.
The problem we're trying to solve is that hotel cancellations can be costly for both hotels and guests. By predicting cancellations in advance, hotels can better manage their inventory and minimize revenue loss, while guests can avoid cancellation fees.